mysite/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.56.1
django-countries==7.5.1
django-debug-toolbar==4.2.0
django-filter==23.2
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.4
filetype==1.2.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
l18n==2021.3
lxml==4.9.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
parsel==1.8.1
Pillow==10.0.0
pillow-heif==0.13.0
Protego==0.3.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.10.1
selenium==4.12.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tldextract==3.5.0
trio==0.22.2
trio-websocket==0.10.4
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
w3lib==2.1.2
wagtail==5.1.1
wagtail-metadata==5.0.0
wagtailcodeblock==1.29.0.1
webencodings==0.5.1
Willow==1.6.2
wsproto==1.2.0
zope.interface==6.0
